+++ common/go.mod
@@ -4,7 +4,6 @@ go 1.25.0
 
 require (
 	github.com/gardener/scaling-advisor/api v0.0.0
-	github.com/gardener/scaling-advisor/minkapi v0.0.0-20260212140146-b3c4c0032c64
 	github.com/go-logr/logr v1.4.3
 	github.com/go-logr/stdr v1.2.2
 	github.com/google/go-cmp v0.7.0
